fix(gasto-anuncios): review fixes — bigint amount, calendar dates, notes, a11y
Correctness (blockers from the first review):
- amountCents: integer -> bigint (schema, hand-edited migration SQL and
  snapshot for now; both get regenerated once #27 merges and the
  migration is renumbered into the 9xxx range, per instructions).
- Zod: real calendar-date validation ("2026-02-31" -> 400, not a 500 from
  Postgres rejecting an invalid date at insert) and a sane upper bound on
  amountCents.
- wonCentsPorFuente: lte -> lt on the end bound, matching ads.ts's
  convention, plus organization_id on its joins (same fix as the
  feat/resultados review, applied here since this file has the same
  pattern).
- AdSpendSummaryDto gained hasWonWithoutAmount (won deals that count as a
  customer but can't count toward return) and hasOtherCurrencySpend
  (spend in another currency, silently excluded from the totals above);
  both surfaced as a footnote under the 4 cards only when true.

Designer pass (partial — see report for what's not done):
- Empty-period copy is now the same "nadie ha cargado gasto para estas
  fechas" across all 4 cards, not a different sentence per card.
- Spend cards use money's `compact` option now that it has a caller.
- Dialog: role="alert"/aria-live on both field and top-level errors, a
  brief "Carga guardada." confirmation after saving, and delete buttons
  now name the period too, not just the source.
- Added a large-amount case to proratedCents' tests (still well under
  Number.MAX_SAFE_INTEGER).

+++ drizzle/0022_gasto_anuncios.sql
@@ -4,7 +4,7 @@ CREATE TABLE "ad_spend" (
 	"source" text NOT NULL,
 	"period_start" date NOT NULL,
 	"period_end" date NOT NULL,
-	"amount_cents" integer NOT NULL,
+	"amount_cents" bigint NOT NULL,
 	"currency" text NOT NULL,
 	"note" text,
 	"created_by" text,
